Wait for session to exit in unshare_private_domain_test
Context: All cf commands write to the config before exiting

We have a theory that the helper which removes the home dir is
failing because the CF command in this unshare_private_domain_test is
still writing to the temp config inside the home dir when the test
cleanup runs. This happens because if we dont wait for session to exit,
the test code continues while session runs asynchronously.

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Eli Wrenn <ewrenn@pivotal.io>