Do not filter by space when getting route to map
In v7 'map-route', we get a route based on domain guid/hostname/path
and (previously) space guid. When trying to map a route that already
existed in a different space, filtering in this way would mean that we
could not find the route in the current space and would try to create
it, which would fail with a message like "Route already exists".

By changing this to not filter by space, we find the route in the other
space, try to map it, and fail with a more useful message like "Routes
cannot be mapped to destinations in different spaces"

[Finishes #163610807]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>