Makefile: export GOFLAGS depending on target
This allows ginkgo to run in the right context (V7 or V6) depending on
the $TARGET_V7 environment variable when running make commands.

We have refrained from propagating the `GOFLAGS` throughout the Makefile
because it is overridden by the command line. Many of our Make targets
use the `-tags` flag, and that would override our GOFLAGS setting.

Co-authored-by: Luan Santos <lsantos@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>