`cf apply-manifest` has improved error messages
When passing a path to a manifest (or directory) via the `-f` flag, `cf
apply-manifest` gives a richer description of the failure, a description
similar to other `cf` commands.

We introduce a new flag, `ManifestPathWithExistenceCheck`, specifically
designed for the `apply-manifest` subcommand. Not only does it check
that the pathname passed exists, but also ensures that if the pathname
is a directory, that the directory contains a file named
`manifest.{yaml,yml}`

[finishes #166525241]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>