v7: `push` with `-f <dir>` ensures dir contains manifest file
Previously, we only validated that `-f` was a path that existed on the filesystem.
If this were a directory that didn't contain a manifest, we would end up getting
a confusing error message like either:

`Name must be a string` (if the current directory was not empty)

or

`Application files not found` (if the current directory was empty)

Now, we use the new `ManifestPathWithExistenceCheck` flag type to
do this validation early and report an incorrect-usage error.

[Finishes #166851328]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>