Use new flags.V7RoutePath for v7 --path flags
- For route-related commands, we want to accept both "/some-path" and
"some-path". We were doing this normalization (adding a leading slash if
necessary) in several places, but now we are doing it when we unmarshal
the flag values.

- Renamed old v6 RoutePath to V6RoutePath to disambiguate. We could not
use this for v7 commands because it behaves slightly differently (for
empty string values, it still adds a "/", but in v7 we want to leave
these as empty).

- Fixes bug in implementation of 'check-route', which was not
normalizing the --path flag and thus passing "some-path" without a slash
to the API.

[Finishes #166805008]
[Finishes #166144539]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>