update-user-provided-service should not unset values
- Issue https://github.com/cloudfoundry/cli/issues/1694 describes an
incorrect behavour where update-user-provided-service can unset existing
credentials associated with a service even when the -p flags is not
specified. We should fix that so that values are not changed unless the
associated flag is specified. Unsetting a value should be done by
explicitly setting in to an empty string.
- In the process we refactored the update-user-provided-service command
to be cocnsistent with the v6 commands

[finishes #166830864]
[PR]

Co-authored-by: Niki Maslarski <nikolay.maslarski@sap.com>