Users cannot push a Docker manifest-app with `-b` flag
The `-b` flag specifies a buildpack (e.g. `ruby-buildpack`), which is
incompatible with a Docker app. We refuse to push; instead, we give a
helpful warning message, e.g.:

```
Incorrect Usage: The argument --buildpack cannot be used with the manifest property docker
```

The other case is already handled (when passing `-o` when pushing
buildpack manifest):

```
Cannot create Docker package for a buildpack app.
```

[finishes #167524206]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>