Validate use of Docker with path properties/flags combo
[#167526885](https://www.pivotaltracker.com/story/show/167526885)

Previously, one could successfuly push an app that defines a Docker
image in its manifest while also passsing in the `--path` flag. The
resulting app would use the Docker image and ignore the provided path.

Now, for clarity, the CLI will reject the push command when it detects
that these conflicting properties are provided.

Similarly, using the `--docker-image` flag with the `path` manifest
property was changed to validate the usage in the same way.

Also, fixes a panic when push with `--path` or `--buildpacks` and no
manifest (or a manifest without applications).

[Finishes #167575890]
[Finishes #167526885]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>