Fix pipeline error for plugin test GetSpaceUsers
- The integration plugin test was testing the Plugin wrapper command
GetSpaceUsers which relies on the v6 create-space command's
functionality of setting the user as a space developer and space
manager.
- The V7 create-space command does not include that functionality
yet; it will be added to the V7 command once the relevant V3 endpoints
are implemented. We have written out and commented out command layer tests to drive out
implemnentation of that functionality in a future story.
- We got the integration plugin tests to pass by separately using the
set-space-role CLI commands to set the user as a space developer and
manager after space creation for now.

[finishes #167168581]

Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>
Co-authored-by: Harsha Nandiwada <hnandiwada@pivotal.io>