Fix plugin integration tests
On V6, create-org and create-space both assign the current user a few
roles (SpaceManager, SpaceDeveloper for spaces; OrgManager for orgs). On
V7 that's not currently implemented because the V3 endpoints for role
management don't yet exist.

This commit changes the helpers for the plugins suite so that they
separately assign the user the correct roles after org/space creation.
It is also done for V6, but redundantly so, since V6 already takes care
of it as part of creating the resources.

[#167167164]

Co-authored-by: Brendan Smith <brsmith@pivotal.io>