v7: Enforce consistent style for idempotent warnings
For commands where we want to output a warning like "App 'app-name' does
not exist." or "App 'app-name' already exists.", followed by "OK", the
V7 style guide states we should not have an empty line between these
messages. The existing UI.DisplayWarning message added two newlines
after the warning, so we are introducing UI.DisplayWarningV7, which only
includes a single newline. This gives us more flexibility in V7. We did
not change the existing method because we did not want to break V6
output.

For `create-route`, we also changed the warning to be printed to stderr
instead of stdout, to be consistent with other commands.

[Finishes #160467192]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>