Fix up tests for `cf labels` for buildpack
There was some strange behavior with the introduction of the `--stack`
flag with the buildpack resource. There was some specific behavior which
is now better captured in the tests that we love so much.

For example, when given the --stack flag but an "" (empty string) is
given, it is treated as if no --stack flag is given.

This is consistent with the cf-buildpack related commands. We do not
love what it looks like, but we love that it is consistent.

[#167362807](https://www.pivotaltracker.com/story/show/167362807)

Co-authored-by: Steve Taylor <staylor@pivotal.io>