WIP
The approach here is to use concourse tasks and a ruby utility script
from CAPI to replicate how they rotate certs.

Their scripts assume the presence of a cf-deployment, which makes their
approach fail for rotating the Concourse director's credentials.  The
first pass at solving the problem is to steal their scripts, make them
work for us, and later see if we can generalize the solution so that
CAPI and CLI can share the same set of credential rotation scripts.

Here's how this is supposed to work:
1. Get bbl-state
2. Remove certificates from the director's vars store
3. Put bbl-state
4. bbl-up with cf-deployment-concourse-tasks' task

Other thoughts:
1. Why are we still using vars stores for our director?  Can't we use
credhub now?
2. We'll need this for our long-lived integration testsing environments
as well.

[#167980583]

Authored-by: Steve Taylor <staylor@pivotal.io>