Publish RedHat packages for both V6 & V7
We publish both the V6 and the V7 `.rpm` installers to our RedHat repo.
We accomplish this by merging the V6 and V7 installers in a common
directory which we then use publish the repo.

We are forced to use one task (because we pull both in), which makes
things simpler. The job will trigger on either a V6 or V7 build.

[#166481406]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>