Separate jobs for V6 & V7 RedHat packages
We thought we could get away with a single job; we couldn't. The
difficulty lay in that, on the release laptop, one of the inputs would
always be gray, and thus the job would never trigger.

Our solution was to have two almost-identical jobs, one triggered by V6,
the other, V7.

[#166481406]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>