v7: Do not hard-code username in integration tests
Some domain-related integration tests had a username hard-coded in the
expected output, which was failing when the tests were run in
client-credentials mode (since it was a different user executing the
commands). This change grabs the currently-authenticated username and
expects that to appear (as we were already doing in most other
integration tests).

[#168090974]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>