DRY up "Not logged in" text
We saw similar warnings in the `command` directory. This behavior is
still being tested through the unit tests
`command/v6/api_command_test.go` and `command/v6/login_command_test.go`.
We believe a separate test for this new file is unnecessary, as it would
be mainly testing the behavior of `go.fmt`.

[#158460421](https://www.pivotaltracker.com/story/show/158460421)

Co-authored-by: Eric Promislow <eric.promislow@suse.com>