Fix CI flake stemming from route being deleted before the service instance
The culprit was peculiar behaviour of fakeservicebroker helper when
there are left-over un-purged service instances from previous usage
of the reusable broker.

Because we’re creating a new config for reusable broker, we don’t know
what was the service offering name for that existing instance, as we
are generating a new random name here.

The fix is to purge existing service instances not based on the config
of a fake service broker, but based on its catalog instead, which would
always purge real, currently existing service offerings.

[finishes #168383869]
[PR]

Co-authored-by: Oleksii Fedorov <ofedorov@pivotal.io>