Users can give relative paths with the `-p` flag
We were treating the `-p` flag as a manifest value so if it was relative
we would treat it as relative to the manfiest and not the directory
where the push was run. We now just always expand any path flag to be an
absolute path in the UnmarshalFlag function of PathWithExistenceCheck.

[finishes #168545851]

Co-authored-by: Eli Wrenn <ewrenn@pivotal.io>
Co-authored-by: Supraja Narasimhan <snarasimhan@pivotal.io>