Refactor V6 apply_test file to avoid depending on real time
This is intended to fix pipeline flakes that we have seen a lot
recently. This file was using a helper method to read events from
streams incrementally, waiting for the event that was being asserted on,
and using a timeout in case that event never appeared within some amount
of time.

Since these tests were relying on real time, they were prone to being
flaky. Instead, we read all events from all streams into arrays, then
assert on the contents of the array.

To keep the expectations structured similarly to how they were, we
introduced two new matchers:

- ContainElementsInOrder, which looks for the
given elements in an array in the specified order but ignores other
elements
- ContainElementTimes, which looks for the given element a specified
number of times in an array

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Eli Wrenn <ewrenn@pivotal.io>