Use the time to get a more random appName
- This was causing two apps in different spaces to have the same names
so when we tried to push, they got the same routes causing CAPI to error
saying routes can't have destinations in a space other than the route.
We now use the time the push command was executed to make the names
random.

- We could not use the helper NewAppName() function because the name +
random word combo would be too long for a hostname

Co-authored-by: Eli Wrenn <ewrenn@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>