Make duplicate-app v3 error message more permissive
We recently changed the API to improve the error message when trying to
create an app with a duplicate name. Now, it looks like "App with the
name 'eli' already exists.", but previously it looked like "name must be
unique in space".

Since we want this to be an idempotent case on the `create-app`/`rename`
commands, we needed to be able to match the error message and catch it
without exiting with a nonzero exit code. To preserve this behavior in
newer clients talking to older APIs, we are expanding the string
matching to look for either the old error message or the new one.

This is especially important because changing the exit code accidentally
would be a breaking change.

[#168654360]

Co-authored-by: Eli Wrenn <ewrenn@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>