Environment variables must be strings
Our old tests used numbers (not strings), but we've made a decision to
conform with CAPI's interface, which only accepts strings, rather than
taking on the burden of converting integers to strings on-the-fly.

We remove the tests that previously set integer variables.

[#167954123]

fixes:
```
displays the environment variables [BeforeEach]
/Users/pivotal/go/src/code.cloudfoundry.org/cli/integration/v7/global/env_command_test.go:110

No future change is possible.  Bailing out early after 0.454s.
Expected
    <int>: 1
to match exit code:
    <int>: 0
```

Co-authored-by: Michelle He <mhe@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>