v7: Implement `cf set-running-environment-variable-group`
Similar to `cf set-staging-environment-variable-group`, allows users to
update the staging environment variables.

In doing this, found and fixed a bug in the implementation of
`set-staging...`: we were handling it correctly when users set new
environment variables (or updated existing ones), and when users cleared
all environment variables by passing `{}`. But we were not handling
"partial unsets" correctly:

If I had set `{"key1": "val1", "key2": "val2"}`, then `{"key1":
"val1"}`, the expected behavior is to unset `key2`. This wasn't working,
but should work now.

[Finishes #167954128]
[Fixes #167953997]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>