v7: Implement `cf set-space-role` command
Allows setting space role for a user. This command has the same flag
options as `cf set-org-role.` Not that to assign a space role, a user
must be a member of organization for that space. This commands creates
an `organization_user` role for the user before creating the space role.

[finishes #164451046]

Co-authored-by: Supraja Narasimhan <snarasimhan@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>