run-task accepts a process to execute (#1817)
add --process flag. will use the supplied process as a template for
the task. the process command, memory, disk will be used to execute a
one-off task.

remove COMMAND as a required positional argument

add --command flag. shorthand -c. optional method to supply the command to execute for
the task.

if neither --process nor --command flags are provided, then assume there is a
process named 'task' and attempt to execute it.

addresses the run-task updates for Issue #1641