v7: Update set-{org,space}-role global integration tests
We made some changes on the API side to improve the error message when
the user does not exist. Also, the commands now look for the org/space
by name before trying to do anything with the user, so we have to make
sure the org/space actually exist in our test. This gets past the
"org/space not found" error to let us really test the user-not-found
case.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>