`cf logs` uses v7
Unlike most commands, the `cf logs` doesn't hit many CAPI endpoints;
rather, it hits the log endpoints, which can be found on the root, (i.e.
GET /). Here is a redacted version of the logging links. Note that the
`logging` link (doppler) is deprecated, and instead we use the newer
`log_cache` interfaces:

```json
{
   "links": {
      "logging": {
         "href": "wss://doppler.iris-eater.lite.cli.fun:443"
      },
      "log_cache": {
         "href": "https://log-cache.iris-eater.lite.cli.fun"
      }
   }
}
```

Also, use a goroutine to periodically refresh the AccessToken
(like the `cf log tail` people do).  If no other cf commands
are updating the config's AccessToken, then the logCacheClient's
request to the log server will be refused and it will stop working.

Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Eric Promislow <eric.promislow@suse.com>
Co-authored-by: Supraja Narasimhan <snarasimhan@pivotal.io>

[finishes #168654370](https://www.pivotaltracker.com/story/show/168654370)