Split generic BuildpackInvalidMessage from API (revised from Reverted commit)
- This code currently indicates two different situations:
  - Buildpack exists with stack mismatch
  - Proposed buildpack name is invalid

We now store the original message on the error object
(which we renamed to more accurately reflect the error).
We differentiate it in the command handler when the message is
emitted.

These only affect the create and rename commands.  Invalid names
don't apply for update, and the stack conflict for updates is handled
by a different error code.

[#168183422](https://www.pivotaltracker.com/story/show/168183422)

Authored-by: Eric Promislow <eric.promislow@suse.com>