Use helper method to assert on staging log lines
The exact log lines that appear when we stream logs during staging is
not always guaranteed. We were making a series of assertions on the
output that were too strict, and thus the test was failing a fraction of
the time because some lines were not appearing.

This change replaces those lines with a call to
`helpers.ConfirmStagingLogs`, which was an existing helper method that
makes a more flexible assertion on the session output.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>