v7: Fix `org-users` and `space-users` commands
Fixes a few small issues:

- For `org-users`, when the `--all-users` flag was passed, before this
change you would see duplicate users listed if a user had multiple roles
in the org. This change de-dupes the users when listing all org users.

- For the empty states of these tables ("No <ROLE> found"), we were
displaying the role name "with flavor", so it would appear blue. We have
moved away from the pattern in v7, so removing the flavor here.

[Finishes #164451030, #164450995]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>