Honor windows installer previously selected install path
In #128697009 we changed the windows installer to set the install path
dependent on the whether the user selected install for "Anyone who uses
this computer" or "Only for me".

This had the side-effect that when the user attempted to run the
installer again their previously selected install path was ignored.

Instead we now check the installation path and use a custom path if the
user specified one earlier.

Also:

- We now always show the directory that we are installing to
  (AlwaysShowDirOnReadyPage)
- We considered setting DisableDirPage to "no" to allow the user to
  always specify the installation directory (even on upgrade).
  We chose not to do this because we were concerned that this would
  allow us to get into other states such as having multiple copies of
  the CLI on the filesystem.
- DefaultDirName is now set to "{pf}\Cloud Foundry" (with a space) for
  consistency with the NextButtonClick logic which specifies that path
  with a space.

[#149572253](https://www.pivotaltracker.com/story/show/149572253)

Co-authored-by: Alexander Berezovsky <aberezovsky@pivotal.io>