🐞 `delete-user` properly checks user is deleted
Prior behavior: `delete-user` says it successfully deleted the user even
when CC has not deleted the user.

This was a side-effect of delete-user not checking the job to make sure
it succeeded.

We now check the job, and, if the job fails, we return the error &
indicate the user was not deleted.

[finishes #170007083]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>