v7: Use random org/space for `set-space-role` integration tests
Previously, we were using `ReadOnlyOrg` and `ReadOnlySpace`, but we were
doing more than just reading: we were creating roles that were
associated with these orgs and spaces. Now, we generate random org/space
names and create them before each test.

This might fix the failures we have been seeing in the
client-credentials tests in the pipeline, but it might have no effect.

[#169967549]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>