🐞 V7 Integration: clarify failing test
This test was unclear. Specifically, it re-used the UAA client
(`potato-face`) in a manner that polluted the test.

- The client (`potato-face`) created an org and a space and a user, a
side-effect being that the client was now a member of the org and space.

- The logged-in user was changed to the newly-created user, which was a
space manager.

- As a space manager, the test attempted to assign a role to the client
(`potato-face`); we wanted it to fail, but because it already had a role
in the space, it succeeded.

Our fix was to use a previously unaffiliated UAA client,
`cf_smoke_tests`, which allowed the tests to pass.

Outstanding Mysteries
------

We never determined why the test began to fail when it did; the commits
on the API side and the CLI side were innocuous & unrelated as far as we
could tell.

fixes: <https://ci.cli.fun/teams/main/pipelines/V7/jobs/integration-linux-client-creds/builds/272>
```
  No future change is possible.  Bailing out early after 0.447s.
  Got stuck at:
      Assigning role SpaceAuditor to user potato-face in org INTEGRATION-ORG-1d9b9464-4225-46a1-7838-e1ff72bc8785 / space INTEGRATION-SPACE-53431c9a-1095-4815-4bfc-b41bbcaa7a62 as INTEGRATION-USER-542c7b2a-ce76-43a2-4786-6f00b8092354...
      OK

  Waiting for:
      FAILED

  /tmp/build/06eff510/go/src/code.cloudfoundry.org/cli/integration/v7/isolated/set_space_role_command_test.go:98

```

[finishes #169967549]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>