Add --wait, -w flag to run-task
When the wait flag is supplied to run-task, the cli will
not exit until the task enters a terminal state; SUCCEEDED or FAILED.

The cli exits 0 when the task succeeds and exits 1 when the task fails.

This improves the scriptability of tasks. They can be run
serially in a script.

Addresses Issue #1104

Authored-by: Zach Robinson <zrobinson@pivotal.io>