Rename 'manifestparser' package to 'v6manifestparser'
This version of the manifest parser is currently only used in two
commands: `v3-apply-manifest` in V6, and `apply-manifest` in V7. In a
follow-up commit, we will refactor V7 `apply-manifest` to use a newer
implementation of the manifest parser (currently called
`pushmanifestparser`, but soon to be renamed to something more generic).
At that point, `v6manifestparser` will be used only in V6, and then it
can be removed when all V6 code is removed.

[#170163478]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>