Rename 'pushmanifestparser' to 'manifestparser'
As part of the push refactor a few months ago, we introduced a new and
improved manifest parser package that we called `pushmanifestparser`, to
disambiguate it from the existing `manifestparser` package. That
refactor only touched the `push` command, but we intended to refactor
the V7 `apply-manifest` command to also use this new implementation.

Now, in preparation for that refactor, we are renaming
`pushmanifestparser` to a more general (hopefully permanent) name. We
renamed the existing `manifestparser` to `v6manifestparser` in a
previous commit to make it clear that it was old and should only be used
in V6 code.

[#170163478]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>