`unset-space-role` and `unset-org-role`
- We implemented the two commands in the title

- We add three new CloudControllerClient methods:
  - `GetUser()`, which maps to `v3/users/:guid`, returns one user or
    error
  - `GetUsers()` which maps to `v3/users`, accepts a list of queries
    and returns a list
  - `DeleteRole()` which maps to `-X DELETE v3/roles/:guid`, and is the
    core method for unsetting roles

We wrapped the error returned from cloudcontroller-`ResourceNotFound`
as a CC error `UserNotFound`, and propagate an ActionError
`UserNotFound` which is actually just a rename of the existing
`UAAUserNotFound` that contains the username and origin of user.

This is a monolithic commit that finishes two stories; we'd normally
split these into separate commits, but there was significant overlap,
especially with errors.

[finishes #164451074]
[finishes #164451089]

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Eric Promislow <epromislow@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Supraja Narasimhan <snarasimhan@pivotal.io>