Delete vacuous test.
- Upgrading to golang 1.13, we noticed that this test was attempting to
create a timeout by communicating with an unreachable ip address.  We
noticed this because now the `net` package responds with a more helpful
"network unreachable" error, causing this test to fail.
- We refactored the code to actually time out by pointing at
www.pivotal.io:81.
- We thought more about it, and removed setting CF_DIAL_TIMEOUT, and
guess what?  That passes!  Because it times out.
- We decided that testing that we see a network timeout occur when
there's a network timeout is unhelpful: that test belongs in golang's
network stack.

[#168846870](https://www.pivotaltracker.com/story/show/168846870)