v7 Plugins: GetOrg
Decided to not copy the Actor's object over directly. Instead, we
provide a simpler Organization object. We were dissatisfied with the
`v7action.OrganizationSummary` and did not want to expose unnecessary
fields.

We performed user research on the list of plugins. We looked at the
source code for all plugins published since Jan 2018, and saw that only
one plugin was using `GetOrg`. This plugin was using it for the Org GUID
and the Spaces' names and GUIDs.

[#168848156](https://www.pivotaltracker.com/story/show/168848156)