v7: Update role integration tests
For set-*-role, the API now supports the creation of user resources when
setting a role by GUID. This means that when the `--client` flag is
given, if the user doesn't exist in CAPI's database, it'll be created.

For unset-*-role, the code is currently implemented to be idempotent
when the user does not exist (and thus there is no role to delete).
Updating the tests to reflect that.

[#170197571]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>