v7: Check if client is valid before setting org role
We hit the UAA /oauth/clients/:client_id endpoint to check if the client
is valid, just like we did in V6.

In `set-space-role`, we "swallow" any user-not-found errors that occur while trying
to create an org user role behind-the-scenes. We let the forbidden error
be surfaced because a user would to reauthenticate with clients.read
scope in their accessToken before even being able to add an org
user.

[#170197571]

Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>