Remove V7 suffix on DisplayWarning(s) methods
In the `ui` package, we had four similar methods:
- `DisplayWarning`
- `DisplayWarnings`
- `DisplayWarningV7`
- `DisplayWarningsV7`

The main difference in the V7 methods were that they did not append an
extra newline to the end of the output.

With this change, instead of having all four of these methods always
defined, we use the `V7` build flag to conditionally compile the
versions of the `DisplayWarning`/`DisplayWarnings` methods that we want
to use for each version. All usages of the V7 methods had the V7 suffix
removed.

[Finishes #169106926]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Supraja Narasimhan <snarasimhan@pivotal.io>