v7: Refactor `routes` command to use inline route destinations
Previously, we were making an additional API request for each route in
the list to fetch its destinations. In
https://github.com/cloudfoundry/cloud_controller_ng/commit/e05c3edc327348bd5d9ab874a5c4ddb7576e4387,
we made a change on the API side to include destinations on route
resource objects. This means that the CLI does not need to make an
additional request to fetch destinations.

[Finishes #170631588]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>