v7: Sort 'newest' packages by created_at instead of updated_at
When restaging, we get the 'newest' package that's in the READY state.
Before this change, we were ordering by `updated_at`, which worked fine.
However, if (in the future) metadata updates on packages bump the
`updated_at` timestamp, then the ordering of these packages would change
in potentially unexpected ways.

This change switches to using the `created_at` field for ordering, which
should be stable and more future-proof.

[Finishes #170631100]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>