🐞 CI: V7 pipeline: use correct path to script
[#167390121]

fixes <https://ci.cli.fun/teams/main/pipelines/V7/jobs/integration-windows/builds/1038>:
```
cli/ci/cli-v7/tasks/integration-windows.ps1 : The term 'cli/ci/cli-v7/tasks/integration-windows.ps1' is not recognized
as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was
included, verify that the path is correct and try again.
```

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>