v7: Clean up start command and integration tests
- When polling the app, waiting for it to start, we had an idempotency
check for a user-not-found error. As far as we can tell, there's no way
for this method to actually return this error (we're not fetching
users), so it seems like a copy/paste leftover. Going to get rid of this
check.
- We had some commented-out expectations in the integration tests that
had notes to bring them back when v7 `map-route` is done. Now that it's
done, bringing those back.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>