v7: Implement `space-quota` command
This code is very similar to the code for the new `org-quota` command.
To make things more reusable (e.g. the `QuotaDisplayer`) for both
`SpaceQuota`s and `OrganizationQuota`s, I changed those two structs to
have an embedded `Quota` struct which contains the common fields between
the two types (most of the fields!).

[#165135194]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>