`Command.Setup()` refactored into base class
This commit extracts shared setup into a base class for v7 commands. The
`push` command will be refactored in a later commit.

As a byproduct of the refactor, we introduced a new method to retrieve
the Log Cache endpoint.

It is worth noting that we consolidated all of our actor interfaces into one
interface. We feel this more accurately reflects how we use the actor as
one abstracted unit in the command layer. This also helps reduce the
amount of repeated code.

[#171869110]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Belinda Liu <bliu@pivotal.io>