Fix null-int unmarshalling to work with Go 1.14
From the Go 1.14 release notes:

> Number no longer accepts invalid numbers, to follow the documented behavior more closely. If a program needs to accept invalid numbers like the empty string, consider wrapping the type with Unmarshaler.

See here: https://golang.org/doc/go1.14#encoding/json

Upgrading to Go 1.14 caused a handful of unit tests to fail. These were
tests where we were trying to unmarshal `""` as a number. Before Go
1.14, this unmarshalling would not return an error, but would just
return an empty string indicating that the value was not parsed as a
number. Now, they return a real error.

We now account for this, and we did some refactoring of these similar
unmarshal methods.

Co-authored-by: Nick Guerette <nguerette@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>