Update custom-lint to remove alphabetical ordering
We had a historic custom lint program that would check methods and
functions were in alphabetical order - only in the api and util
packages. We have decided to remove this because it was causing a lot of
CI breaks, and it's not obvious that it makes sense to understand the
code better.

Co-authored-by: George Blue <gblue@pivotal.io>