v7: user can see output while waiting for an app to start
While polling process instances, if an instance has any details on it,
we will report it like "Error starting instance: '<details>'". If not,
we will report a generic error message: "Instances starting...".

We acknowledge that this may be verbose, especially when you have an app
with a larger number of instances. We may revisit this in the future
(e.g. only outputting info when we have real instance details to
report).

We chose to implement this with a "callback" function to allow different
commands to use their own display logic for instance details, rather
than something more complex (e.g. passing in a channel that would
receive instance details, to be read and displayed at the command
layer).

[finishes #171632362]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>